python/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
sqlalchemy==2.0.30
prophet==1.1.5
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
numpy==1.26.4
python-dotenv==1.0.1
